LINQ Dice Simulation
A C# Windows Form Application dice roll simulations demonstrating the
use of LINQ and IEnumerable.Except. The application will roll two dice
100 times and save the results to a list. The application will then
repeat the process and allow the user to use LINQ in order to query the
results of both roll sequences and display the amount of rolls resulting
in a specific total. IEnumerable.Except then allows the user to display
the rolls in sequence one that were not rolled in sequence two.